requirements.txt
streamlit
together
pdfplumber
python-docx
pytesseract
Pillow
pdf2image